login havent logged in thing added

+++ base/views.py
@@ -5,7 +5,9 @@ from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login,logout
+from django.http import HttpResponse
 # Create your views here.
 #function / classes deals with what something people will see if they go to certain url
 
@@ -16,6 +18,10 @@ from django.contrib.auth import authenticate, login,logout
 #]
 
 def loginPage(request):
+    page='login'
+    if request.user.is_authenticated: #if user already logged in, return them to homepage
+        return redirect('home')
+
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')# receiving data from frontend
@@ -31,8 +37,16 @@ def loginPage(request):
         else:
             messages.error(request,'username or password does not exist')
 
-    context={}
+    context={'page':page}
     return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page='register'
+    return render(request,'base/login_register.html')
 
 def home(request):
     q= request.GET.get('q') if request.GET.get('q')!= None else '' # q is whatever we passed through te url if nothing passed in , just set as empty
@@ -55,6 +69,7 @@ def room(request,pk): #can pass in dynamic value
     context={'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='Login') #if user is not authenticated, user will be directed to a page to login
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -66,9 +81,13 @@ def createRoom(request):
     context={'form':form} #pass empty dictionary
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='Login') #if user is not authenticated, user will be directed to a page to login
 def updateRoom(request,pk): #passing pk to know what item we are updating
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room) #fill with the room value 
+
+    if request.user!=room.host: #if the creator of the room is not equal to the guy trying to delete it,
+        return HttpResponse('you are not allowed here')# say that they are not allowed to do so
 
     if request.method=='POST':
         form= RoomForm(request.POST,instance=room) #specifies which room to edit
@@ -79,8 +98,11 @@ def updateRoom(request,pk): #passing pk to know what item we are updating
     context={'form':form} #want to prefill the info
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='Login') #if user is not authenticated, user will be directed to a page to login
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host: #if the creator of the room is not equal to the guy trying to delete it,
+        return HttpResponse('you are not allowed here')# say that they are not allowed to do so
     if request.method=='POST':
         room.delete()
         return redirect('home')
